http://issues.apache.org/activemq/browse/AMQCPP-25
Fix VC compile errors

git-svn-id: https://svn.apache.org/repos/asf/incubator/activemq/activemq-cpp/trunk/activemq-cpp@494990 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/activemq/io/DataOutputStreamTest.h
@@ -41,9 +41,9 @@ namespace io{
 		void test(){
 			
 			unsigned char byteVal = (unsigned char)'T';
-			uint16_t shortVal = 5;
-			uint32_t intVal = 10000;
-			uint64_t longVal = 1000000000;
+			unsigned short shortVal = 5;
+			unsigned int intVal = 10000;
+			unsigned long long longVal = 1000000000;
 			float floatVal = 10.0f;
 			double doubleVal = 100.0;
 			unsigned char arrayVal[3] = {
